\name{fasting}
\alias{fasting}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Metabolome analysis of mouse liver samples in fasting conditions}
\description{
This data includes metabolome data of liver samples from normal and 12 hour fasted mice measured by capillary electrophoresis time-of-flight mass spectrometry. The data matrix contains 10 samples (5 samples in each groups) and 282 metabolites.
}
\usage{
data(fasting)
}
%- maybe also 'usage' for other objects documented here.
\format{
The list object "fasting" contains the following elements:

  X : 	Data matrix that include metabolites in each columns.

  Y :   Dummy matrix that include group information 0,1 in each columns.
}
\source{
Yamamoto H., Fujimori T., Sato H., Ishikawa G., Kami K., Ohashi Y. (2014). "Statistical hypothesis testing of factor loading in principal component analysis and its application to metabolite set enrichment analysis". BMC Bioinformatics, (2014) 15(1):51.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ datasets }

